+++ source/ProcessManager.Orchestrations/Processes/BRS_021/Shared/ElectricityMarket/MeteringPointReceiversProvider.cs
@@ -13,16 +13,14 @@
 // limitations under the License.
 
 using System.Diagnostics.CodeAnalysis;
-using System.Globalization;
 using Energinet.DataHub.ProcessManager.Abstractions.Core.ValueObjects;
 using Energinet.DataHub.ProcessManager.Components.Abstractions.ValueObjects;
-using Energinet.DataHub.ProcessManager.Orchestrations.Abstractions.Processes.BRS_021.ForwardMeteredData.V1.Model;
-using Energinet.DataHub.ProcessManager.Orchestrations.Processes.BRS_021.ForwardMeteredData.V1.Extensions;
-using Energinet.DataHub.ProcessManager.Orchestrations.Processes.BRS_021.ForwardMeteredData.V1.Model;
+using Energinet.DataHub.ProcessManager.Core.Domain.OrchestrationInstance;
+using Energinet.DataHub.ProcessManager.Orchestrations.Processes.BRS_021.Shared.ElectricityMarket.Model;
 using NodaTime;
 using NodaTime.Extensions;
 
-namespace Energinet.DataHub.ProcessManager.Orchestrations.Processes.BRS_021.ForwardMeteredData.ElectricityMarket;
+namespace Energinet.DataHub.ProcessManager.Orchestrations.Processes.BRS_021.Shared.ElectricityMarket;
 
 [SuppressMessage(
     "StyleCop.CSharp.ReadabilityRules",
@@ -40,46 +38,23 @@ public class MeteringPointReceiversProvider(
 
     private readonly DateTimeZone _dateTimeZone = dateTimeZone;
 
-    public List<ReceiversWithMeteredDataV1> GetReceiversWithMeteredDataFromMasterDataList(
-        IReadOnlyCollection<MeteringPointMasterData> meteringPointMasterDataList,
-        ForwardMeteredDataInputV1 input)
+    public List<ReceiversWithMeasureData> GetReceiversWithMeteredDataFromMasterDataList(
+        FindReceiversInput input)
     {
-        if (meteringPointMasterDataList.Count == 0)
+        if (input.MasterData.Count == 0)
             throw new InvalidOperationException($"The metering point master data list is empty (MeteringPointId={input.MeteringPointId}, StartDateTime={input.StartDateTime}, EndDateTime={input.EndDateTime})");
 
         // Ensure metered data is sorted by position
-        var sortedMeteredData = new SortedDictionary<int, ReceiversWithMeteredDataV1.AcceptedMeteredData>(
-            input.MeteredDataList.ToDictionary(
-                md => int.Parse(md.Position!),
-                md =>
-                {
-                    // TODO: Shouldn't position be an int? in the input?
-                    var position = int.Parse(md.Position!);
-
-                    // TODO: Shouldn't EnergyQuantity be a decimal? in the input?
-                    var canParseEnergyQuantity = decimal.TryParse(
-                        md.EnergyQuantity,
-                        CultureInfo.InvariantCulture,
-                        out var energyQuantity);
-
-                    // The input is already validated, so converting these should not fail.
-                    return new ReceiversWithMeteredDataV1.AcceptedMeteredData(
-                        Position: position,
-                        EnergyQuantity: canParseEnergyQuantity ? energyQuantity : null,
-                        QuantityQuality: Quality.FromNameOrDefault(md.QuantityQuality));
-                }));
+        var sortedMeteredData = new SortedDictionary<int, ReceiversWithMeasureData.MeasureData>(
+            input.MeasureData.ToDictionary(md => md.Position));
 
         // Ensure master data is sorted by ValidFrom
-        var masterDataDictionary = meteringPointMasterDataList.ToDictionary(mpmd => mpmd.ValidFrom.ToInstant());
-
-        // The input is already validated, so this parsing should never fail
-        var inputPeriodStart = InstantPatternWithOptionalSeconds.Parse(input.StartDateTime).Value;
-        var inputPeriodEnd = InstantPatternWithOptionalSeconds.Parse(input.EndDateTime!).Value;
+        var masterDataDictionary = input.MasterData.ToDictionary(mpmd => mpmd.ValidFrom.ToInstant());
 
         var allReceivers = CalculateReceiversWithMeteredDataForMasterDataPeriods(
-            inputPeriodStart,
-            inputPeriodEnd,
-            Resolution.FromName(input.Resolution!), // Resolution shouldn't change between master data periods, else validation should fail
+            input.StartDateTime,
+            input.EndDateTime,
+            input.Resolution, // Resolution shouldn't change between master data periods, else validation should fail
             masterDataDictionary,
             sortedMeteredData);
 
@@ -93,12 +68,12 @@ public class MeteringPointReceiversProvider(
     /// validation should ensure that the resolution is the same for each master data period (in the same transaction).
     /// </remarks>
     /// </summary>
-    private List<ReceiversWithMeteredDataV1> CalculateReceiversWithMeteredDataForMasterDataPeriods(
+    private List<ReceiversWithMeasureData> CalculateReceiversWithMeteredDataForMasterDataPeriods(
         Instant inputPeriodStart,
         Instant inputPeriodEnd,
         Resolution resolution,
         Dictionary<Instant, MeteringPointMasterData> masterData,
-        SortedDictionary<int, ReceiversWithMeteredDataV1.AcceptedMeteredData> sortedMeteredData)
+        SortedDictionary<int, ReceiversWithMeasureData.MeasureData> sortedMeteredData)
     {
         if (masterData.Count == 1)
         {
@@ -153,8 +128,8 @@ public class MeteringPointReceiversProvider(
                 throw new InvalidOperationException($"The current timestamp is equal to or after the master data period end (MeteringPointId={currentMasterData.MasterData.MeteringPointId.Value}, Position={meteredData.Position}, CurrentTimestamp={currentTimestamp}, MasterDataValidTo={currentMasterData.ValidTo})");
 
             // Position is 1-indexed, so the new position is the current count + 1 (if the list is empty, the new position should be 1)
-            var newPosition = currentMasterData.MeteredDataList.Count + 1;
-            currentMasterData.MeteredDataList.Add(meteredData with { Position = newPosition });
+            var newPosition = currentMasterData.MeasureData.Count + 1;
+            currentMasterData.MeasureData.Add(meteredData with { Position = newPosition });
 
             // Get next timestamp
             currentTimestamp = AddResolutionToTimestamp(currentTimestamp, resolution);
@@ -187,24 +162,24 @@ public class MeteringPointReceiversProvider(
         return nextTimestamp;
     }
 
-    private ReceiversWithMeteredDataV1 CreateReceiversWithMeteredData(
+    private ReceiversWithMeasureData CreateReceiversWithMeteredData(
         MasterDataWithMeteredData masterDataWithMeteredData)
     {
         var actorReceivers = GetReceiversFromMasterData(masterDataWithMeteredData.MasterData);
 
-        return new ReceiversWithMeteredDataV1(
-            Actors: actorReceivers,
+        return new ReceiversWithMeasureData(
+            Receivers: actorReceivers,
             Resolution: masterDataWithMeteredData.MasterData.Resolution,
             MeasureUnit: masterDataWithMeteredData.MasterData.MeasurementUnit,
             StartDateTime: masterDataWithMeteredData.ValidFrom.ToDateTimeOffset(),
             EndDateTime: masterDataWithMeteredData.ValidTo.ToDateTimeOffset(),
-            MeteredData: masterDataWithMeteredData.MeteredDataList);
+            MeasureDataList: masterDataWithMeteredData.MeasureData);
     }
 
-    private List<MarketActorRecipientV1> GetReceiversFromMasterData(
+    private List<Actor> GetReceiversFromMasterData(
         MeteringPointMasterData meteringPointMasterData)
     {
-        var receivers = new List<MarketActorRecipientV1>();
+        var receivers = new List<Actor>();
         var meteringPointType = meteringPointMasterData.MeteringPointType;
 
         switch (meteringPointType)
@@ -222,7 +197,7 @@ public class MeteringPointReceiversProvider(
                 receivers.AddRange(
                     meteringPointMasterData.NeighborGridAreaOwners
                         .Select(ActorNumber.Create)
-                        .Select(NeighborGridAccessProviderReceiver));
+                        .Select(GridAccessProviderReceiver));
                 break;
             case var _ when meteringPointType == MeteringPointType.VeProduction:
                 receivers.Add(SystemOperatorReceiver());
@@ -234,6 +209,20 @@ public class MeteringPointReceiversProvider(
                 }
 
                 break;
+
+            // Electrical heating, net consumption and capacity settlement metering points always sends to the energy supplier and grid access provider
+            case var _ when meteringPointType == MeteringPointType.ElectricalHeating:
+            case var _ when meteringPointType == MeteringPointType.NetConsumption:
+            case var _ when meteringPointType == MeteringPointType.CapacitySettlement:
+
+                // There can be periods where no energy supplier is assigned to the parent/child metering point,
+                // thus we can only send to the energy supplier if there actually is one.
+                if (meteringPointMasterData.EnergySupplier is not null)
+                    receivers.Add(EnergySupplierReceiver(meteringPointMasterData.EnergySupplier));
+
+                receivers.Add(GridAccessProviderReceiver(meteringPointMasterData.GridAccessProvider));
+                break;
+
             case var _ when meteringPointType == MeteringPointType.NetProduction:
             case var _ when meteringPointType == MeteringPointType.SupplyToGrid:
             case var _ when meteringPointType == MeteringPointType.ConsumptionFromGrid:
@@ -275,26 +264,34 @@ public class MeteringPointReceiversProvider(
         }
 
         var distinctReceivers = receivers
-            .DistinctBy(r => (r.ActorNumber.Value, r.ActorRole.Name))
+            .DistinctBy(r => (r.Number.Value, r.Role.Name))
             .ToList();
 
         return distinctReceivers;
     }
 
-    private MarketActorRecipientV1 EnergySupplierReceiver(ActorNumber energySupplierId) =>
+    private Actor EnergySupplierReceiver(ActorNumber energySupplierId) =>
         new(energySupplierId, ActorRole.EnergySupplier);
 
-    private MarketActorRecipientV1 NeighborGridAccessProviderReceiver(ActorNumber neighborGridAccessProviderId) => new(
-        neighborGridAccessProviderId,
+    private Actor GridAccessProviderReceiver(ActorNumber gridAccessProviderId) => new(
+        gridAccessProviderId,
         ActorRole.GridAccessProvider);
 
-    private MarketActorRecipientV1 DanishEnergyAgencyReceiver() => new(
+    private Actor DanishEnergyAgencyReceiver() => new(
         ActorNumber.Create(DataHubDetails.DanishEnergyAgencyNumber),
         ActorRole.DanishEnergyAgency);
 
-    private MarketActorRecipientV1 SystemOperatorReceiver() => new(
+    private Actor SystemOperatorReceiver() => new(
         ActorNumber.Create(DataHubDetails.SystemOperatorNumber),
         ActorRole.SystemOperator);
+
+    public sealed record FindReceiversInput(
+        string MeteringPointId,
+        Instant StartDateTime,
+        Instant EndDateTime,
+        Resolution Resolution,
+        IReadOnlyCollection<MeteringPointMasterData> MasterData,
+        IReadOnlyCollection<ReceiversWithMeasureData.MeasureData> MeasureData);
 
     private sealed record MasterDataWithMeteredData
     {
@@ -302,12 +299,12 @@ public class MeteringPointReceiversProvider(
             MeteringPointMasterData masterData,
             Instant inputPeriodStart,
             Instant inputPeriodEnd,
-            List<ReceiversWithMeteredDataV1.AcceptedMeteredData> meteredDataList)
+            List<ReceiversWithMeasureData.MeasureData> measureData)
         {
             MasterData = masterData;
             ValidFrom = Instant.Max(inputPeriodStart, masterData.ValidFrom.ToInstant());
             ValidTo = Instant.Min(inputPeriodEnd, masterData.ValidTo.ToInstant());
-            MeteredDataList = meteredDataList;
+            MeasureData = measureData;
         }
 
         public MeteringPointMasterData MasterData { get; }
@@ -316,6 +313,6 @@ public class MeteringPointReceiversProvider(
 
         public Instant ValidTo { get; }
 
-        public List<ReceiversWithMeteredDataV1.AcceptedMeteredData> MeteredDataList { get; }
+        public List<ReceiversWithMeasureData.MeasureData> MeasureData { get; }
     }
 }